using WriterModules enum where possible
