+++ nexus_constructor/json/forwarder_json_writer.py
@@ -4,8 +4,9 @@ import h5py
 
 from nexus_constructor.json.helpers import object_to_json_file
 from nexus_constructor.nexus.nexus_wrapper import get_nx_class
+from nexus_constructor.writer_modules import WriterModules
 
-FORWARDER_SCHEMAS = ["f142", "TdcTime"]
+FORWARDER_SCHEMAS = [WriterModules.F142, WriterModules.TDCTIME]
 
 
 def find_forwarder_streams(root, provider_type: str) -> List:
