+++ nexus_constructor/writer_modules.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class WriterModules(Enum):
+    F142 = "f142"
+    EV42 = "ev42"
+    TDCTIME = "TdcTime"
+    NS10 = "ns10"
+    HS00 = "hs00"
+    SENV = "senv"
